Boddat Fixes
Fixed problems with creating bva vector and searching for cases with PM, Pole, and Eq.

In getnum function, kept orientation of input bodies.

In mkeph, fixed issues with not reading ephemeris data correctly and creating temp variables that are the correct type.

In putsbmb, fixed the issues with adding to asteroids or major bodies beyond current allocated space. Fixed issues when finding pp (5 vector).